Fused class, header and SealModule.cc file.
https://github.com/cms-sw/cmssw/pull/31722#discussion_r603659618
