--- RecoPixelVertexing/PixelTrackFitting/plugins/SealModule.cc
@@ -1,5 +1,0 @@
-#include "FWCore/PluginManager/interface/ModuleDef.h"
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "PixelTrackProducer.h"
-DEFINE_FWK_MODULE(PixelTrackProducer);
